Move PoolRow to own file
